git push delete function works if you are the creator of the post

+++ public/js/deletePost.js
@@ -3,17 +3,19 @@ const $deleteBtn = document.querySelector('#delete_post');
 async function deletePost(e) {
     e.preventDefault();
     try {
-        const id = window.location.toString().split('/')[
-            window.location.toString().split('/').length - 1
-        ];
-        const response = await fetch(`/myposts/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json'
-            },
-        });
-        location.href = "/myposts"
+        if (confirm('Are you sure that you want to delete this post?') == true) {
+            const id = window.location.toString().split('/')[
+                window.location.toString().split('/').length - 1
+            ];
+            const response = await fetch(`/myposts/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    'Accept': 'application/json',
+                    'Content-Type': 'application/json'
+                },
+            });
+            location.href = "/myposts";
+        }
     } catch (error) {
         $error.textContent = error;
         console.log(error);
